Updated arrows so that they are not centered on the ship transform, and instead rotate their position around the ship in addition to rotating direction.
Shots will now fire directly towards the other ship regardless of whether your ship is centered on it, on the conditions that you are within a specifiable distance of the other ship and that the angle between your ship's facingdirection vector and the vector between your ship and the other ship is less than a certain degree.

Hid the bullseye.

Cargo now respawns after 5 seconds instead of after 1.

Sped up ships a little bit.