controllers should work now
The controllers should work for both the mac and pc controllers now.
Controls are:
boost = A
break = B
shoot = left or right bumpers

Also, I changed up the left and right joysticks. Let me know how you all
like it. If not enough controllers are plugged in, the default is to use
the keyboard for the rest. @duncsmit @pramodsum @EvanHahn